moves TestType and random_known_rpc_port_config to test_type.rs and config.rs
